requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.3.0
filelock==3.19.1
filetype==1.2.0
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
groq==0.32.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.7
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.6
mpmath==1.3.0
narwhals==2.6.0
networkx==3.4.2
numpy==2.2.6
ollama==0.6.0
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.3.3
pi_heif==1.1.1
pillow==11.3.0
pillow-avif-plugin==1.5.2
polars==1.33.1
protobuf==6.32.1
psutil==7.1.0
pyarrow==21.0.0
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
roboflow==1.2.9
rpds-py==0.27.1
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.50.0
supervision==0.26.1
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.204
ultralytics-thop==2.0.17
urllib3==2.5.0
